RobustPade

Documentation for RobustPade.

RobustPade.robustpade

RobustPade.robustpade

RobustPade.robustpade

RobustPade.robustpade

RobustPade.robustpade_size

RobustPade.robustpade_size

RobustPade.robustpade_table

RobustPade.robustpade — Method

robustpade(
    f::Polynomial,
    m::Integer,
    n::Integer;
    kwargs...
)

computes the (m,n) Pade approximant corresponding to the Taylor expansion given by a f::Polynomials.Polynomial

RobustPade.robustpade — Method

robustpade(
    f::Taylor1,
    m::Integer,
    n::Integer;
    kwargs...
)

computes the (m,n) Pade approximant corresponding to the Taylor expansion given by a f::TaylorSeries.Taylor1

RobustPade.robustpade — Method

robustpade(
    coeffs::AbstractVector{T},
    m::Integer,
    n::Integer;
    tol::Real=eps(float(real(T)))
) where T

computes the (m,n) Pade approximant to a function with Taylor coefficients coeffs::AbstractVector using SVD following Parchon et al. SIAM Review.

Adapted from the Chebfun implementation at https://github.com/chebfun/chebfun/blob/master/padeapprox.m (modified BSD license)

RobustPade.robustpade — Method

robustpade(
    f::Function,
    m::Integer,
    n::Integer,
    x::T=0.;
    tol::Real=100eps(float(real(T)))
) where T<:Number

computes the (m,n) Pade approximant to a function f using TaylorSeries.taylor_expand to compute the Taylor expansion at x.

RobustPade.robustpade_size — Method

robustpade_size(f::AbstractVector,m::Integer,n::Integer;kwargs...)

computes the degrees of the (m,n) Pade approximant corresponding to the Taylor expansion coefficients f

RobustPade.robustpade_size — Method

robustpade_table(f,m::Integer,n::Integer,args...;kwargs...)

computes the degrees of the (m,n) Pade approximant corresponding to f

RobustPade.robustpade_table — Method

robustpade_table(f,M::Integer,N::Integer,args...;kwargs...)

computes the (0:M,0:N) Pade table corresponding to f